Introduce user pool security profiles, refactor tests
