+++ backend/compact-connect/cdk.context.sandbox-example.json
@@ -10,6 +10,7 @@
         "region": "us-east-1",
         "domain_name": "justin.compactconnect.org",
         "allow_local_ui": true,
+        "security_profile": "VULNERABLE",
         "notifications": {
           "ses_operations_support_email": "justin@example.com",
           "email": [
