+++ backend/compact-connect/stacks/persistent_stack/provider_users.py
@@ -31,7 +31,7 @@ class ProviderUsers(UserPool):
             removal_policy=removal_policy,
             email=user_pool_email,
             sign_in_aliases=None,
-            standard_attributes=_configure_user_pool_standard_attributes(),
+            standard_attributes=self._configure_user_pool_standard_attributes(),
             custom_attributes={
                 # these fields are required in order for the provider to be able to query
                 # for their personal profile information. Once these fields are set, they cannot be changed.
@@ -67,89 +67,90 @@ class ProviderUsers(UserPool):
         )
 
 
-def _configure_user_pool_standard_attributes() -> StandardAttributes:
-    """
-        The provider user pool standard attributes.
+    @staticmethod
+    def _configure_user_pool_standard_attributes() -> StandardAttributes:
+        """
+            The provider user pool standard attributes.
 
-        Note that these values can never change! If you need to make other attributes
-        required in the future, you must create an entirely new user pool and migrate
-        existing users to the new pool. See https://repost.aws/knowledge-center/cognito-change-user-pool-attributes
+            Note that these values can never change! If you need to make other attributes
+            required in the future, you must create an entirely new user pool and migrate
+            existing users to the new pool. See https://repost.aws/knowledge-center/cognito-change-user-pool-attributes
 
-        These attributes are used to display on a provider's profile page. We do not
-        intend to use them for authentication purposes or for back-end processing.
-    """
-    return StandardAttributes(
-        # We are requiring the following attributes for all users
-        # that are registered in the provider user pool.
-        email=StandardAttribute(
-            mutable=True,
-            required=True
-        ),
-        given_name=StandardAttribute(
-            mutable=True,
-            required=True
-        ),
-        family_name=StandardAttribute(
-            mutable=True,
-            required=True
-        ),
-        # The following attributes are not required, but we are including them because
-        # Cognito does not allow you to add them after the user pool is created, and we
-        # may want to use them in the future.
-        address=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        birthdate=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        fullname=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        gender=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        last_update_time=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        locale=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        middle_name=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        nickname=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        phone_number=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        preferred_username=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        profile_page=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        profile_picture=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        timezone=StandardAttribute(
-            mutable=True,
-            required=False
-        ),
-        website=StandardAttribute(
-            mutable=True,
-            required=False
+            These attributes are used to display on a provider's profile page. We do not
+            intend to use them for authentication purposes or for back-end processing.
+        """
+        return StandardAttributes(
+            # We are requiring the following attributes for all users
+            # that are registered in the provider user pool.
+            email=StandardAttribute(
+                mutable=True,
+                required=True
+            ),
+            given_name=StandardAttribute(
+                mutable=True,
+                required=True
+            ),
+            family_name=StandardAttribute(
+                mutable=True,
+                required=True
+            ),
+            # The following attributes are not required, but we are including them because
+            # Cognito does not allow you to add them after the user pool is created, and we
+            # may want to use them in the future.
+            address=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            birthdate=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            fullname=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            gender=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            last_update_time=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            locale=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            middle_name=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            nickname=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            phone_number=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            preferred_username=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            profile_page=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            profile_picture=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            timezone=StandardAttribute(
+                mutable=True,
+                required=False
+            ),
+            website=StandardAttribute(
+                mutable=True,
+                required=False
+            )
         )
-    )
